requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
decorator==5.1.1
diskcache==5.6.1
EasyProcess==1.1
elevenlabs==0.2.24
entrypoint2==1.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.101.1
ffmpeg==1.4
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
iniconfig==2.0.0
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
keyboard==0.13.5
loguru==0.7.0
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numpy==1.24.3
openai==0.27.6
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
pluggy==1.2.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.4.0
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.15.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
pytest==7.3.1
python-dotenv==1.0.0
pytweening==1.0.7
PyYAML==6.0.1
regex==2023.8.8
requests==2.30.0
rich==13.3.5
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
simpleaudio==1.0.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.32.0
typing_extensions==4.7.1
urllib3==2.0.2
uvicorn==0.23.2
wcwidth==0.2.6
websockets==11.0.3
win32-setctime==1.1.0
yarl==1.9.2
